implement question.picture while playing quiz
