+++ frontend/src/components/pages/sessions/QuestionDisplay.tsx
@@ -1,5 +1,6 @@
 import { AnswerButtons } from "@/components/pages/sessions/AnswerButtons";
 import { RankingModal } from "@/components/pages/sessions/RankingModal";
+import SupabaseImage from "@/components/shared/SupabaseImage";
 import { QUIZ_STATES } from "@/constants/quizState";
 import { useQuiz } from "@/stores/quizStore";
 import type { Question } from "@/types/Question";
@@ -51,7 +52,23 @@ export function QuestionDisplay({
         <div className="w-full mt-2 mb-6">
           <TimerContainer showResults={showResults} onTimeExpire={handleTimeExpire} />
         </div>
-        <h2 className="grow text-xl font-bold mt-2 mb-6 text-center">{question.questionText}</h2>
+
+        {question.picture ? (
+          <>
+            <div className="mb-4">
+              <SupabaseImage
+                fileName={question.picture}
+                width={300}
+                height={200}
+                alt="Question Image"
+              />
+            </div>
+            <h2 className="text-xl font-bold mb-6 text-center">{question.questionText}</h2>
+          </>
+        ) : (
+          <h2 className="grow text-xl font-bold mt-2 mb-6 text-center">{question.questionText}</h2>
+        )}
+
         <AnswerButtons
           question={question}
           onAnswer={(optionId) => {
@@ -67,7 +84,7 @@ export function QuestionDisplay({
             </>
           )}
         </div>
-      </div>
+      </div>{" "}
       {quizState === QUIZ_STATES.RESULTS && showResults && (
         <RankingModal open={showRankingModal} questionTotal={questionTotal} />
       )}
